Fixed the SecurityTicket release status matcher. Added valid status config.
It was resulting in strange status values, mainly because there was no
test suite for these details (and the ticket that was being tested didn't
had the "(rXXXXX)" comments, so it didn't triggered the problem).

In order to properly match the status lines, it required to define which
were the valid status entries, so it was added one configuration option to
define what is valid. It also required to make SecurityTicket and
TicketSource objects receive the "config" objects as parameter.