pull-cves: Store CVE objects instead of raw XML
As we already spend some time downloading the XML file, we can afford
converting these raw XMLs to CVE objects at the same time. Changed the
format from bsddb + rawstrings to shelve.

Also introduced a file "crontabs.txt" suggesting the commands to be used to
pull all the data that is needed to be cached locally.