Implemented a package pool
This package pool is intended to keep some metadata about all available
packages on the distro. It should help in the tasks of working in the
security updates and also in finding relevant CVEs in the CVE database.

For now the base is imeplemented in sqlite and all distro medias should be
referenced explicitly.