Introduced more commands to search in the kernel changelog
Introduced --kreleases to list all kernel releases.

Introduced --kcommits to list all commits from a given kernel release.

And finally, introduced --kcve/--krel, which is used to look for evidence
of CVEs in a given kernel releases. It tries to perform all the trickery
needed to check if a commit in a give release fixes a CVE issue.