Moved the CVE database to sqlite
The current implementation was really fast to dump CVE descriptions
quickly, but it was horribly slow to navigate through all CVEs and to parse
the YAML format. The current implementation of pyyaml is too slow.

The initial database model kept CVE references in a separated table, in
order to follow the classic normalization recommendations. The problem is
that it slowed down the process of dumping a CVE, because it required
accessing the database twice to create a CVE object. Thus, it was decided
to have a 'multivalored' field with the references.